Update Sample code for the Hybrid Machine Learning Model
We developed a hybrid approach to address the inaccuracy of quantitative predictions in zero-inflated data with Gaussian distribution. This approach comprises two steps combining qualitative and quantitative models sequentially. In the first step, we transformed the continuous outcome data into a binary variable, distinguishing between zero (zero observations) and one (non-zero observations) values. Then, fit a qualitative model to predict the presence or absence of the outcome. In the second step, we fitted the quantitative model and predicted the continuous outcome in all locations where the qualitative model indicated the presence of the outcome. To fit the model in the second step we only utilized the non-zero observations from the initial continuous data. Following this hybrid approach, the qualitative model served as a tool for identifying where the outcome was likely to occur, and the quantitative model was used to estimate the magnitude of the outcome. The hybrid approach can be viewed as a complementary framework that combines the strengths of the qualitative and quantitative models. The qualitative model excels in handling data with numerous zeros but lacks quantitative capabilities, whereas the quantitative model is adept at quantification but can struggle with datasets containing many zero values.